move codes to GameViewModel
